Updating handler API and moving them to own folder
