+++ tests/test_communication.py
@@ -3,7 +3,7 @@ import unittest
 from simulator.event import EventLoop
 from simulator.messages.communication import CommunicationCommand, CommunicationCommandType
 from simulator.node import Node
-from simulator.node.communication import Medium, can_transmit, CommunicationHandler, CommunicationException
+from simulator.node.handler.communication import Medium, can_transmit, CommunicationHandler, CommunicationException
 
 
 def handle_command_helper(command: CommunicationCommand):
